+++ spectral_cube/tests/test_dask.py
@@ -106,8 +106,8 @@ def test_statistics(data_adv):
 
 def test_statistics_withnans(data_adv):
     cube = DaskSpectralCube.read(data_adv).rechunk(chunks=(1, 2, 3))
-    # shape is 2, 3, 4
-    cube._data[:,:,:2] = np.nan
+    # shape is 4, 3, 2 for adv
+    cube._data[:2,:,:] = np.nan
     # ensure some chunks are all nan
     cube.rechunk((1,2,2))
     stats = cube.statistics()
@@ -266,7 +266,11 @@ def test_cube_on_cube(filename, request):
     # since they are not SpectralCube subclasses
     cube = DaskSpectralCube.read(dataname)
     assert isinstance(cube, (DaskSpectralCube, DaskVaryingResolutionSpectralCube))
-    cube2 = SpectralCube.read(dataname, use_dask=False)
+    if 'image' in filename:
+        # no choice - non-dask can't read casa images
+        cube2 = SpectralCube.read(dataname)
+    else:
+        cube2 = SpectralCube.read(dataname, use_dask=False)
     if 'image' not in filename:
         # 'image' would be CASA and must be dask
         assert not isinstance(cube2, (DaskSpectralCube, DaskVaryingResolutionSpectralCube))